Initial add of database stages
These stages of the WordPress database allow us to quickly set WordPress up for specific scenarios. The first is an initial installation of WordPress in single site mode. The second is an installation in multisite mode with only one site. These can be copied to {project-dir}/database/backups/wsuwp.sql to be automatically imported into the environment on initial vagrant up.