Start implementing a new structure for the WordPress admin bar
Adds a new menu item for 'My Networks' and separates and moves some of the existing functionality of 'My Sites' to cause less confusion for users that may be members of multiple sites. The left area of the screen should just work as if it was a single site. Additional advanced options will be available on the right.