Include MP6 as an initial plugin
The styles provided by MP6 will most likely be included with WordPress core by the time 3.8 is released. Including the plugin in the repository now will allow us to plan for that release.

This would be better handled by a submodule, but a git mirror is not yet availble for the MP6 project. Once it is included in core, this will most likely be removed.