Reorganize the menu items in the WordPress admin bar to better suit our needs
Places our My Networks item all the way to the left and then keeps the standard order for the other items just to the right of that. Some additional logic will soon be needed to only display the My Networks item when a user is a member of more than one network.