Modify admin menu behavior based on role and number of sites
If the user is not a super admin and only has access to one site, then we should avoid confusion by hiding the My Sites menu item and not showing the My WSU Networks menu item.

If the user is not a super admin, but has access to multiple sites, then we will show the standard My Sites menu item, but not worry about showing any WSU networks.

In the future, we'll need to determine the correct roles and capabilities to use rather than the blanket is_super_admin approach.