Introduce wp_get_user_networks, or at least part of it
We'd like to know what networks the user has access to. It seems that we're able to do this by looking at the sites they have access to and then pulling off the network IDs with wp_list_pluck to do another lookup for the full network objects.

Leaving this hanging as some easy fruit to come back to next week.