Introduce switch_to_network() and restore_current_network()
This allows us to to switch between networks when doing things like displaying details of networks while showing a user's available networks in the admin bar.

I would assume very unstable at this point. We should come back and cache this info and see what else might be missing or might be needed in order for things to work properly.