Show only the user's networks in the My Networks menu
Previously, we continued to show all sites that the user was a member of rather than just networks. This change makes use of the new wp_get_user_networks() and switch_to_network() to fill the admin menu with relevant information.

Next we'll add a sub menu for each of the networks that lists all of the user's sites on those networks.