Redirect invalid subdomains via sunrise.php
If an invalid subdomain for any of our networks is detected after a site cannot be found using the domain and path to look it up, then redirect to what we can assume is the "parent" subdomain by removing the first portion of the domain.

If we do this **and** see that the site is not one that we are aware of, then redirect to the domain specified as the master domain in wp-config.php

We'll need to do some verification of this behavior as we go.