Implement object cache handling of site routing
Take advantage of our object cache when querying for awareness of a specific site on the network. Stores and gets values for the site itself and for the network that the site belongs to. Theoretically the network will not need to be queried for often once it is in cache.

We may need to consider a model where we store only the site ID for the domain/path combo as it's possible many key/value stores will be created for paths that do not actually have an individual site.