Progress on network creation - several new functions introduced
We're taking over the functionality normally offered by populate_network and using a new function to fill in network data as well as all of the associated meta data. In the process, we're going a bit overboard by providing functions to add and update individual network meta options. These aren't fleshed out at all, but may come in handy.