Add an action link for global plugin activation on the main network
If we're looking at the plugins for the main network, we should have the option to globally activate or deactivate something. We should make sure that an option for globally deactivating is available as well.

It'd be interesting if there was a kill switch that disabled a plugin across the network even if it wasn't a globally activated plugin.