Introduce is_plugin_active_for_global()
Introduces new function to determine if the plugin has been activated globally and uses this when decided to display the 'Globally Activate' text under the plugin.

Also sneaks in a change to handle the globally active plugins through the global wp_options table rather than the table of the main site.