Replacing php54w-mysql with php54w-common for now...
I think I'm going to move to another CentOS repository, as this one seems a bit light on the modules that we should account for. I'd like to use Imagick for processing images rather than GD and the PECL RPM doesn't appear to be avaialble here. This should be the first change tomorrow.