Add provisioning to update the existing wsuwp-platform git repository
If things are already checked out, then we'll want to keep them updated.

This may lead to mapping the /var/tmp/wsuwp-platform directory to the local machine when working with the VM so that these files persist. On the production server, the files will remain in tmp at all times... likely.