Modify yum configuration before Salt provisioning
By default, the bootstrap script run by Salt to install itself on the server detects that this is a CentOS machine and then issues a `yum -y update`. This makes every package on the system fair game for an automatic upgrade, including the kernel. Something about the kernel upgrade that I don't understand yet causes the Virtualbox guest additions to disappear, leaving us without a bootable VM unless we destroy it each time.

By excluding the kernel in yum.conf, we're able to speed up the bootstrap process significantly *and* avoid the issues it was causing.

This commit does cause a few other things to happen out of order for some reason. It now takes 2 provisions to get everything actually working. The next work will take care of that.