Don't provide a persistent database after `vagrant destroy`
I've reversed on this a bit. We should enforce good workflow rather than trying to account for absolute convenience, especially when the persistence of this database can cause issues.

Use `vagrant halt` or `vagrant suspend` to avoid losing DB changes.
Use `vagrant destroy` to wipe the entire DB clean.

We'll want to provide some default DB setup files for when it is wiped clean.