Map persistent www/ directory to /var/www/ in development environment
Rather than relying on a pseudo deployment script in our local development environment from rsync, we should just map the drive with persistent data.

This does introduce a really hacky scenario. By default, there is no www-data user or group in the CentOS box that we're using, so we don't have the option to mount /var/www before provisioning starts unless we're okay with the Vagrant user. We allow that to happen and then unmount and remount the drive as www-data once the primary provisioning script runs.

Now that I better understand the issue, there's a chance that we can move the unmount/mount into the wsuwp-dev SLS for a cleaner Vagrantfile.