Remove wsuwp-dev-update to allow for responsible workflow
Rather than make the decision to update the repository every time provisioning is run, we should leave it to the developer to manage the git workflow and check for updates. This would have also caused annoyances if work on other branches was being done or if there was a dirty working tree.