Mount /var/www with a different user if the machine exists
During initial provisioning, the www-data is not available. We mount as the vagrant user and then unmount/mount during provisioning with the newly created www-data when it comes time.

During `vagrant up` after `vagrant halt`, drives are remounted but provisioning is not run. Before, this would cause /var/www/ to be moutned as the vagrant user without the benefit of our additional provisioning.

With this commit, we use `machine_exists` to add www-data owner data when we know that the machine has already been provisioned once (hopefully) and has the www-data user available.