Add WSU Nameservers rather than Google's nameservers.
It would be interesting to see how this could be abstracted out. I'd rather use closer DNS servers (WSU) if possible, but if another university starts using the project, additional work would need to be done to maintain their own fork with their own nameservers.

It would be fun to find a way to make this more configurable in the future via pillars or something similar.