Remove currently unused WSUWP Vagrant configuration
This Vagrant configuration was project specific to WSUWP. The current objective of this repository is to provide provisioning for the WSU Web Communication servers. At some point, a series of helpful Vagrant configurations could live here or in other repositories.

The Vagrant configuration we are removing can be found at WSUWP Platform on GitHub