Remove (currently) unnecessary host management in state file
The method we use here isn't lovely, Salt allows for something much better.

We also take advantage of vagrant-hosts to do this for us inside the guest machine, so this line is not needed.

In the future, it could be smart to handle host management with Salt in a different fashion.