Remove ngx_pagespeed module from nginx build script
We don't take advantage of this module in production on the WSUWP
platform and barely use it or sometimes have it disabled on the
indie environment for only one site. Rather than have it sit as a
possible source for bugs, we should remove it until we know we're
using it with intent.