Provide a config for innodb_buffer_pool_instances
This is a default of 1 in MySQL already. On servers that have an
innodb buffer pool of more than 1GB it can make sense to have
multiple instances so that concurrent threads can process requests.

On WSUWP, we already have a buffer pool of 4GB (soon 6GB) that can
likely benefit from this setting.